read_idats

read_idats

read_idats(directory = getwd(), csv_file = "csv$",
  meth_value_type = "B", filter_beads = TRUE, bead_cutoff = 0.05,
  filter_non_cpg = TRUE, filter_snps = TRUE, population = NULL,
  filter_multihit = TRUE, filter_xy = TRUE, detection_pvalues = 0.01,
  filter_callrate = TRUE, callrate_samples = 0.99,
  callrate_probes = 1, norm_background = "oob", norm_dye = "RELIC",
  norm_quantile = "quantile1", array_name = c("EPIC", "450k"),
  annotation_version = c("ilm10b4.hg19", "ilmn12.hg19"), n_cores = 1)

Arguments

directory
A character. Location of IDAT files, default is the current working directory.

csv_file
A character. Path to the sample sheet (csv files) or name of the sample sheet in directory.

meth_value_type
A character. Indicates whether you prefer m-values ("M") or beta-values ("B"). Default is "B".

filter_beads
A logical. Should probes with a beadcount less than 3 will be removed depending on the bead_cutoff value. Default is TRUE.

bead_cutoff
A numeric. The beadCutoff represents the fraction of samples that must have a beadcount less than 3 before the probe is removed.(default = 0.05)

filter_non_cpg
A logical. Should non-cg probes be removed? Default is TRUE.

filter_snps
A logical. Should probes in which the probed CpG falls near a SNP be removed? Default is TRUE.

population
A vector(character).

filter_multihit
A logical. Should probes in which the probe aligns to multiple locations with bwa be removed? Default is TRUE.

filter_xy
A logical. Should probes from X and Y chromosomes be removed? Default is TRUE.

detection_pvalues
A numeric.

filter_callrate
A logical.,

callrate_samples
A numeric.

callrate_probes
A numeric.

norm_background
A character.

norm_dye
A character.

norm_quantile
A character.

array_name
A character. Choose microarray type, eiyther "450K" or "EPIC". Default is "EPIC".

annotation_version
A character. Version of the annotation package that should be used. Default is "ilm10b4.hg19" for the "EPIC" array

n_cores
An integer. The number of cores to use, i.e., at most how many child processes will be run simultaneously.

Value

A list.